use strongly-type TRE fields
